MySql10.java: Added cast

+++ AccesoDatos/DataBases/MySql10/src/MySql10.java
@@ -33,7 +33,7 @@ public class MySql10 {
 			llamadaProcedimiento.registerOutParameter(2, Types.FLOAT);
 			llamadaProcedimiento.registerOutParameter(3, Types.FLOAT);
 			llamadaProcedimiento.executeUpdate();
-			System.out.println("Salario medio de los empleados: "+llamadaProcedimiento.getFloat(2)+"\n Cantidad de empleados: "+llamadaProcedimiento.getFloat(3));
+			System.out.println("Salario medio de los empleados: "+llamadaProcedimiento.getFloat(2)+"\n Cantidad de empleados: "+(int)llamadaProcedimiento.getFloat(3));
 			
 		} catch (NumberFormatException e) {
 			// TODO: handle exception
